(Commit 1 by Oriol)
Started the whole project from zero to improve it.
Created and added the Java Old Plain Objects Forecast, Day & Location.
Created the methods to get the information from the JSON files on the API Openweathermap.
Created the methods to get the images from the JSON files on the API Openweathermap.
Added menu options for:
-Forecast X city.
-Selected city details.
-Forecast X days.
-Degrees (Celsius/Fahrenheit).
-Average temperature.
-Help Dialog.
-Close.

Everything is working correctly.